Bug 1858542 - Add Glean metrics for sponsored Firefox Suggest impressions.
This commit builds on bug 1857092 to record impressions for sponsored
Firefox Suggest suggestions. It adds a new value for the
`fx_suggest.ping_type` metric, and Desktop's `position`,
`suggested_index`, and `is_clicked` metrics, and sends these new metrics
for both impressions and clicks.

An impression means, "did the user see a suggestion when they finished
engaging with the awesomebar?" A "completed engagement" means they
navigated to either a URL, a search results page, or a shown suggestion.

We don't record impressions for:

* Sponsored suggestions that the user sees as they're typing.
* Abandoned engagements, when the user dismisses the awesomebar without
  navigating to a destination.

If the user taps on a sponsored suggestion, we send an impression ping
_and_ a click ping for that suggestion. If they tap on any other
suggestion, or navigate to a URL or search results page, we send an
impression ping only.

This commit also moves the click ping logic from `MetricController`
to `MetricsMiddleware`. The `Facts` plumbing in the `feature-fxsuggest`
component doesn't know the `position` or `suggested_index` of the
suggestion in the awesomebar, because `onSuggestionClicked` doesn't
have access to that information. Since `MetricsMiddleware` needs to
know whether the sponsored suggestion was the one that was clicked,
it can also be responsible for sending the click ping.